Refactor CSGesture code to common file. Use common Mouse Wrapper for all CSGesture devices. Use common csgesture file for Cypress/Elan Trackpads.
