+++ VoodooI2C/VoodooCSGestureMouseWrapper.cpp
@@ -6,54 +6,42 @@
 //  Copyright © 2015 Alexandre Daoud. All rights reserved.
 //
 
-#include "VoodooCyapaMouseWrapper.h"
-#include "VoodooCyapaGen3Device.h"
+#include "VoodooCSGestureMouseWrapper.h"
+#include "csgesture.h"
 
-OSDefineMetaClassAndStructors(VoodooCyapaMouseWrapper, IOHIDDevice)
+OSDefineMetaClassAndStructors(VoodooCSGestureMouseWrapper, IOHIDDevice)
 
-static VoodooI2CCyapaGen3Device* GetOwner(const IOService *us)
-{
-    IOService *prov = us->getProvider();
-    
-    if (prov == NULL)
-        return NULL;
-    return OSDynamicCast(VoodooI2CCyapaGen3Device, prov);
-}
-
-bool VoodooCyapaMouseWrapper::start(IOService *provider) {
-    if (OSDynamicCast(VoodooI2CCyapaGen3Device, provider) == NULL)
-        return false;
-
+bool VoodooCSGestureMouseWrapper::start(IOService *provider) {
     IOLog("VoodooI2C: %s, line %d\n", __FILE__, __LINE__);
     setProperty("HIDDefaultBehavior", OSString::withCString("Trackpad"));
     return IOHIDDevice::start(provider);
 }
 
-IOReturn VoodooCyapaMouseWrapper::setProperties(OSObject *properties) {
+IOReturn VoodooCSGestureMouseWrapper::setProperties(OSObject *properties) {
     IOLog("VoodooI2C: %s, line %d\n", __FILE__, __LINE__);
     return kIOReturnUnsupported;
 }
 
-IOReturn VoodooCyapaMouseWrapper::newReportDescriptor(IOMemoryDescriptor **descriptor) const {
+IOReturn VoodooCSGestureMouseWrapper::newReportDescriptor(IOMemoryDescriptor **descriptor) const {
     IOLog("VoodooI2C: %s, line %d\n", __FILE__, __LINE__);
-    IOBufferMemoryDescriptor *buffer = IOBufferMemoryDescriptor::inTaskWithOptions(kernel_task, 0, GetOwner(this)->reportDescriptorLength());
+    IOBufferMemoryDescriptor *buffer = IOBufferMemoryDescriptor::inTaskWithOptions(kernel_task, 0, gestureEngine->reportDescriptorLength());
 
     if (buffer == NULL) return kIOReturnNoResources;
-    GetOwner(this)->write_report_descriptor_to_buffer(buffer);
+    gestureEngine->write_report_descriptor_to_buffer(buffer);
     *descriptor = buffer;
     IOLog("VoodooI2C: %s, line %d\n", __FILE__, __LINE__);
     return kIOReturnSuccess;
 }
 
-IOReturn VoodooCyapaMouseWrapper::setReport(IOMemoryDescriptor *report, IOHIDReportType reportType, IOOptionBits options) {
+IOReturn VoodooCSGestureMouseWrapper::setReport(IOMemoryDescriptor *report, IOHIDReportType reportType, IOOptionBits options) {
     IOLog("VoodooI2C: %s, line %d\n", __FILE__, __LINE__);
     return kIOReturnUnsupported;
 }
 
-IOReturn VoodooCyapaMouseWrapper::getReport(IOMemoryDescriptor *report, IOHIDReportType reportType, IOOptionBits options) {
+IOReturn VoodooCSGestureMouseWrapper::getReport(IOMemoryDescriptor *report, IOHIDReportType reportType, IOOptionBits options) {
     IOLog("VoodooI2C: %s, line %d\n", __FILE__, __LINE__);
     if (reportType == kIOHIDReportTypeOutput){
-        GetOwner(this)->write_report_to_buffer(report);
+        gestureEngine->write_report_to_buffer(report);
         return kIOReturnSuccess;
     }
     return kIOReturnUnsupported;
@@ -66,47 +54,47 @@ IOReturn VoodooCyapaMouseWrapper::getReport(IOMemoryDescriptor *report, IOHIDRep
     return IOHIDDevice::handleReport(report, reportType, options);
 }*/
 
-OSString* VoodooCyapaMouseWrapper::newManufacturerString() const {
+OSString* VoodooCSGestureMouseWrapper::newManufacturerString() const {
     IOLog("VoodooI2C: %s, line %d\n", __FILE__, __LINE__);
-    return OSString::withCString("Cypress");
+    return OSString::withCString("CSGesture");
 }
 
-OSNumber* VoodooCyapaMouseWrapper::newPrimaryUsageNumber() const {
+OSNumber* VoodooCSGestureMouseWrapper::newPrimaryUsageNumber() const {
     IOLog("VoodooI2C: %s, line %d\n", __FILE__, __LINE__);
     return OSNumber::withNumber(kHIDUsage_GD_Mouse, 32);
 }
 
-OSNumber* VoodooCyapaMouseWrapper::newPrimaryUsagePageNumber() const {
+OSNumber* VoodooCSGestureMouseWrapper::newPrimaryUsagePageNumber() const {
     IOLog("VoodooI2C: %s, line %d\n", __FILE__, __LINE__);
     return OSNumber::withNumber(kHIDPage_GenericDesktop, 32);
 }
 
-OSNumber* VoodooCyapaMouseWrapper::newProductIDNumber() const {
+OSNumber* VoodooCSGestureMouseWrapper::newProductIDNumber() const {
     IOLog("VoodooI2C: %s, line %d\n", __FILE__, __LINE__);
-    return OSNumber::withNumber(GetOwner(this)->productID(), 32);
+    return OSNumber::withNumber(gestureEngine->productID, 32);
 }
 
-OSString* VoodooCyapaMouseWrapper::newProductString() const {
+OSString* VoodooCSGestureMouseWrapper::newProductString() const {
     IOLog("VoodooI2C: %s, line %d\n", __FILE__, __LINE__);
-    return OSString::withCString("Gen3 Trackpad");
+    return OSString::withCString("Trackpad");
 }
 
-OSString* VoodooCyapaMouseWrapper::newSerialNumberString() const {
+OSString* VoodooCSGestureMouseWrapper::newSerialNumberString() const {
     IOLog("VoodooI2C: %s, line %d\n", __FILE__, __LINE__);
     return OSString::withCString("1234");
 }
 
-OSString* VoodooCyapaMouseWrapper::newTransportString() const {
+OSString* VoodooCSGestureMouseWrapper::newTransportString() const {
     IOLog("VoodooI2C: %s, line %d\n", __FILE__, __LINE__);
     return OSString::withCString("I2C");
 }
 
-OSNumber* VoodooCyapaMouseWrapper::newVendorIDNumber() const {
+OSNumber* VoodooCSGestureMouseWrapper::newVendorIDNumber() const {
     IOLog("VoodooI2C: %s, line %d\n", __FILE__, __LINE__);
-    return OSNumber::withNumber(GetOwner(this)->vendorID(), 16);
+    return OSNumber::withNumber(gestureEngine->vendorID, 16);
 }
 
-OSNumber* VoodooCyapaMouseWrapper::newLocationIDNumber() const {
+OSNumber* VoodooCSGestureMouseWrapper::newLocationIDNumber() const {
     IOLog("VoodooI2C: %s, line %d\n", __FILE__, __LINE__);
     return OSNumber::withNumber(123, 32);
 }
