+++ VoodooI2C/VoodooElanTouchpadDevice.h
@@ -89,8 +89,6 @@ enum tp_mode {
 };
 
 class VoodooI2C;
-class VoodooElanTouchpadWrapper;
-class IOBufferMemoryDescriptor;
 
 class VoodooI2CElanTouchpadDevice : public VoodooI2CDevice
 {
@@ -98,7 +96,7 @@ class VoodooI2CElanTouchpadDevice : public VoodooI2CDevice
     OSDeclareDefaultStructors(VoodooI2CElanTouchpadDevice);
     
 private:
-    VoodooElanTouchpadWrapper* _wrapper;
+    CSGesture* _wrapper;
     
     void initialize_wrapper(void);
     void destroy_wrapper(void);
@@ -150,12 +148,6 @@ public:
 .registerIndex = offsetof(struct i2c_hid_desc, wCommandRegister)
     };
     
-    struct {
-        UInt8 x;
-        UInt8 y;
-        UInt8 buttonMask;
-    } lastmouse;
-    
     uint8_t prevreport[ETP_MAX_REPORT_LEN];
     
     I2CDevice* hid_device;
@@ -170,14 +162,6 @@ public:
     
     void get_input(OSObject* owner, IOTimerEventSource* sender);
     
-    int reportDescriptorLength();
-    
-    int vendorID();
-    int productID();
-    
-    void write_report_to_buffer(IOMemoryDescriptor *buffer);
-    void write_report_descriptor_to_buffer(IOMemoryDescriptor *buffer);
-    
     IOReturn setPowerState(unsigned long powerState, IOService *whatDevice);
     
     int i2c_get_slave_address(I2CDevice* hid_device);
@@ -189,19 +173,6 @@ public:
     SInt32 readI2C16(uint16_t reg, size_t len, uint8_t *values);
     SInt32 writeI2C(uint8_t reg, size_t len, uint8_t *values);
     SInt32 writeI2C16(uint16_t reg, size_t len, uint8_t *values);
-    
-    void update_relative_mouse(char button,
-                               char x, char y, char wheelPosition, char wheelHPosition);
-    void update_keyboard(uint8_t shiftKeys, uint8_t *keyCodes);
-    
-    int distancesq(int delta_x, int delta_y);
-    bool ProcessMove(csgesture_softc *sc, int abovethreshold, int iToUse[3]);
-    bool ProcessScroll(csgesture_softc *sc, int abovethreshold, int iToUse[3]);
-    bool ProcessThreeFingerSwipe(csgesture_softc *sc, int abovethreshold, int iToUse[3]);
-    
-    void TapToClickOrDrag(csgesture_softc *sc, int button);
-    void ClearTapDrag(csgesture_softc *sc, int i);
-    void ProcessGesture(csgesture_softc *sc);
     void TrackpadRawInput(struct csgesture_softc *sc, uint8_t report[ETP_MAX_REPORT_LEN], int tickinc);
 };
