+++ VoodooI2C/csgesture.h
@@ -4,6 +4,11 @@
 #ifndef csgesturestruct_h
 #define csgesturestruct_h
 
+#include <stdint.h>
+#include <IOKit/IOService.h>
+#include <IOKit/IOLib.h>
+#include "VoodooCSGestureMouseWrapper.h"
+
 #define MAX_FINGERS 15
 
 struct csgesture_softc {
@@ -72,6 +77,44 @@ struct csgesture_softc {
     int truetick[15];
     int ticksincelastrelease;
     int tickssinceclick;
+};
+
+class CSGesture {
+private:
+    VoodooCSGestureMouseWrapper *_wrapper;
+    
+    struct {
+        UInt8 x;
+        UInt8 y;
+        UInt8 buttonMask;
+    } lastmouse;
+    
+    int distancesq(int delta_x, int delta_y);
+    
+    //os callbacks
+    void update_relative_mouse(char button,
+                               char x, char y, char wheelPosition, char wheelHPosition);
+    void update_keyboard(uint8_t shiftKeys, uint8_t *keyCodes);
+public:
+    //public csgesture functions
+    bool ProcessMove(csgesture_softc *sc, int abovethreshold, int iToUse[3]);
+    bool ProcessScroll(csgesture_softc *sc, int abovethreshold, int iToUse[3]);
+    bool ProcessThreeFingerSwipe(csgesture_softc *sc, int abovethreshold, int iToUse[3]);
+    
+    void TapToClickOrDrag(csgesture_softc *sc, int button);
+    void ClearTapDrag(csgesture_softc *sc, int i);
+    void ProcessGesture(csgesture_softc *sc);
+    
+    //os specific functions
+    void initialize_wrapper(IOService *service);
+    void destroy_wrapper(void);
+    
+    int vendorID;
+    int productID;
+    
+    int reportDescriptorLength();
+    void write_report_to_buffer(IOMemoryDescriptor *buffer);
+    void write_report_descriptor_to_buffer(IOMemoryDescriptor *buffer);
 };
 
 #endif
